feat: implement writing from delta minio to clickhouse with exactly batch_version, avoid scanning records in last batch

+++ src/batch_jobs/schema/person_schema.py
@@ -1,4 +1,4 @@
-from pyspark.sql.types import StructType, StructField, ArrayType, StringType, FloatType, LongType, IntegerType
+from pyspark.sql.types import StructType, StructField, ArrayType, StringType, DoubleType, LongType, IntegerType
 
 PERSON_FULL_SCHEMA = StructType([
     StructField("person_id", LongType(), True),
@@ -12,6 +12,6 @@ PERSON_FULL_SCHEMA = StructType([
         StructField("deathday", StringType(), True),
         StructField("place_of_birth", StringType(), True),
         StructField("known_for_department", StringType(), True),
-        StructField("popularity", FloatType(), True),
+        StructField("popularity", DoubleType(), True),
     ]), True)
 ])
